+++ app/Http/Controllers/Admin/Content/FAQController.php
@@ -17,7 +17,8 @@ class FAQController extends Controller
      */
     public function index()
     {
-        $faqs = FAQ::orderBy('created_at')->simplePaginate(15);
+        $this->authorize('index', Faq::class);
+        $faqs = Faq::orderBy('created_at')->simplePaginate(15);
         return view('admin.content.faq.index', compact('faqs'));
 
     }
@@ -29,6 +30,7 @@ class FAQController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Faq::class);
         return view('admin.content.faq.create');
     }
 
@@ -40,6 +42,7 @@ class FAQController extends Controller
      */
     public function store(FAQRequest $request)
     {
+        $this->authorize('create', Faq::class);
         $inputs = $request->all();
         Faq::create($inputs);
         return redirect()->route('admin.content.faq.index')->with('alertify-success', 'سوال جدید با موفقیت اضافه شد.');
@@ -64,6 +67,7 @@ class FAQController extends Controller
      */
     public function edit(Faq $faq)
     {
+        $this->authorize('update', Faq::class);
         return view('admin.content.faq.edit', compact('faq'));
     }
 
@@ -76,6 +80,7 @@ class FAQController extends Controller
      */
     public function update(FAQRequest $request, Faq $faq)
     {
+        $this->authorize('update', Faq::class);
         $inputs = $request->all();
         $faq->update($inputs);
         return redirect()->route('admin.content.faq.index')->with('alertify-warning', 'سوال موردنظر با موفقیت ویرایش شد.');
@@ -88,12 +93,14 @@ class FAQController extends Controller
      */
     public function destroy(Faq $faq)
     {
+        $this->authorize('delete', Faq::class);
         $faq->delete();
         return redirect()->route('admin.content.faq.index')->with('alertify-error', 'سوال موردنظر با موفقیت حذف شد.');
     }
 
     public function status(Faq $faq)
     {
+        $this->authorize('update', Faq::class);
         $faq->status = $faq->status == 0 ? 1 : 0;
         $result = $faq->save();
